requirements.txt
absl-py==1.3.0
aiomysql==0.1.1
alembic==1.8.1
altair==4.2.0
aniso8601==9.0.1
anyio==3.6.2
apipkg==1.5
appdirs==1.4.4
argcomplete==2.0.0
argon2-cffi==21.1.0
asgiref==3.6.0
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
asyncio==3.4.3
attrs==20.2.0
Babel==2.11.0
backcall==0.1.0
Beaker==1.10.0
beautifulsoup4==4.11.1
black==22.10.0
bleach==4.0.0
blinker==1.5
blivet==3.4.4
blivet-gui==2.3.0
Bottleneck==1.3.2
Brlapi==0.8.4
Brotli==1.0.9
cachetools==5.0.0
category-encoders==2.6.0
certifi==2020.6.20
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.11
click==8.0.4
colorama==0.4.4
commonmark==0.9.1
configparser==5.0.1
coverage==6.5.0
crayons==0.4.0
cryptography==36.0.0
cssselect==0.9.2
cupshelpers==1.0
cycler==0.11.0
dasbus==1.6
dash==2.6.2
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dbus-python==1.2.18
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.1
distro==1.6.0
Django==4.1.7
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.2.0
dominate==2.7.0
entrypoints==0.3
et-xmlfile==1.0.1
exceptiongroup==1.0.1
execnet==1.7.1
executing==0.8.2
factory-boy==3.2.1
Faker==15.1.1
fastjsonschema==2.16.2
fedora-third-party==0.10
filelock==3.0.12
flaky==3.7.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-Compress==1.13
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-MySQLdb==1.0.1
flask-restplus==0.13.0
flask-shell-ipython==0.4.1
Flask-SQLAlchemy==3.0.2
flask-swagger==0.2.14
flatbuffers==22.11.23
fonttools==4.29.1
freezegun==1.2.2
fros==1.1
frozendict==2.3.4
fs==2.4.11
fsspec==2022.11.0
gast==0.4.0
gcsfs==0.6.2
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.17.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gpg==1.17.0
graphviz==0.20.1
greenlet==1.1.2
grpcio==1.51.1
gunicorn==20.1.0
h5py==3.7.0
html-testRunner==1.2.1
html5lib==1.1
htmlmin==0.1.12
humanize==3.13.1
idna==2.10
ImageHash==4.3.1
imbalanced-learn==0.10.1
imblearn==0.0
importlib-metadata==2.0.0
iniconfig==1.1.1
ipdb==0.13.9
ipykernel==6.6.1
ipyparallel==8.4.1
ipython==8.3.0
ipython_genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.0.1
jdcal==1.4.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.2.0
json5==0.9.10
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.4
jupyter-core==4.9.1
jupyter-server==1.23.3
jupyterlab==3.5.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.16.3
keras==2.11.0
kiwisolver==1.4.1
lab==7.2
langtable==0.0.61
lazypredict==0.2.12
libclang==14.0.6
libcomps==0.1.18
lightgbm==3.3.4
lxml==4.9.2
Mako==1.1.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.2
missingno==0.5.1
mistune==2.0.4
mizani==0.8.1
mock==4.0.3
multimethod==1.9
multitasking==0.0.11
munkres==1.1.2
myModule==1.0.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.33
mysqlclient==2.1.1
nbclassic==0.4.8
nbclient==0.5.10
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.4
netifaces==0.11.0
networkx==2.8.8
nftables==0.1
nltk==3.8
notebook==6.4.12
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.0.2
odfpy==1.4.1
olefile==0.46
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==21.3
paddle-bfloat==0.1.7
paddlepaddle==2.4.2
palettable==3.3.0
pandas==2.0.1
pandas-datareader==0.10.0
pandas-profiling==3.4.0
pandocfilters==1.5.0
parso==0.8.2
Paste==3.5.0
pathspec==0.10.1
patsy==0.5.3
pbr==5.5.0
pexpect==4.8.0
phik==0.12.2
pickleshare==0.7.5
pid==2.2.3
Pillow==9.1.0
platformdirs==2.5.2
plotly==5.10.0
plotnine==0.10.1
pluggy==0.13.1
ply==3.11
productmd==1.35
prometheus-client==0.13.1
prompt-toolkit==3.0.24
protobuf==3.19.6
psutil==5.9.2
psycopg2==2.9.1
ptyprocess==0.6.0
pure-eval==0.2.2
pwquality==1.4.4
py==1.9.0
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.21.0
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.45.1
pydantic==1.10.2
pydeck==0.8.0
pyenchant==3.2.2
pyflakes==2.5.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.4.0
pykickstart==3.36
Pympler==1.0.1
PyMySQL==1.0.3
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyparted==3.12.0
PyQt5==5.15.6
PyQt5-sip==12.9.0
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.2.0
pytest-checkipdb==1.0.9
pytest-cov==3.0.0
pytest-flakes==4.0.5
pytest-flask==1.2.0
pytest-flask-sqlalchemy==1.0.2
pytest-forked==1.3.0
pytest-freezegun==0.4.2
pytest-html==2.1.1
pytest-html-reporter==0.2.3
pytest-lazy-fixture==0.6.3
pytest-metadata==1.10.0
pytest-mock==3.10.0
pytest-randomly==3.11.0
pytest-rerunfailures==10.2
pytest-xdist==2.1.0
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-meh==0.50
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyudev==0.22.0
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
requests==2.28.2
requests-file==1.5.1
requests-ftp==0.3.1
requests-oauthlib==1.3.0
rich==13.0.1
rpm==4.17.1
rsa==4.8
scikit-learn==1.1.3
scikit-surprise==1.1.3
scipy==1.8.1
scour==0.38.2
seaborn==0.12.1
selenium==3.141.0
selinux==3.3
semver==2.13.0
Send2Trash==1.8.0
sepolicy==3.3
setools==4.4.0
simpleaudio==1.0.4
simplejson==3.18.0
simpleline==1.9.0
six==1.15.0
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
sos==4.3
soupsieve==2.3.1
SQLAlchemy==1.4.46
sqlparse==0.4.3
stack-data==0.1.4
statsmodels==0.13.5
stevedore==3.2.2
streamlit==1.17.0
surprise==0.1
systemd-python==234
tables==3.7.0
tabulate==0.8.10
tangled-up-in-unicode==0.2.0
Tempita==0.5.2
tenacity==8.1.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
terminado==0.12.1
testpath==0.5.0
textblob==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.1
tomli==2.0.1
toolz==0.12.0
tornado==6.1
tqdm==4.64.1
traitlets==5.1.1
txt2tags==3.7
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.25.10
validators==0.20.0
virtualenv==20.0.34
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
visions==0.7.5
visitor==0.1.3
watchdog==2.2.1
wcwidth==0.2.5
webdriver-manager==3.2.2
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wordcloud==1.8.2.2
wrapt==1.14.1
xarray==2022.6.0
xgboost==1.7.3
xlrd==2.0.1
XlsxWriter==3.0.8
xlwt==1.3.0
yfinance==0.2.3
zipp==3.3.0
